Update fid_new.py
Now using command line arguments for normalization. If none are provided, no normalization is performed. Normalization has to be done for the main training loop, but also in the code where the metrics are computed. The FID computation assumeds normalized data, since it needs to map to the input range of [0,255] for the inception network. It prints a warning if it finds that the input data is clearly not normalized