Update fid_new.py
Change the way we generate our fake_batches. Also: make sure that if batch_size is larger than num_metric_samples, we take the smallest one as batch size

use the NumpyPathDataset.batch() to get a batch. Also, add the option of NOT passing a writer if you only want to compute (and return) the metrics, not store them in a tf events file