Update optuna_objective.py
Due to the fact that the call to optuna_override_undefined would overwrite the 'args' object, the second trial would THINK that the first trial's parameters were the ones passed on the command line, and reuse the same values. That way, you're running multiple trials, but all with the same parameters - thats useluss. This issue was solved in this commit by making sure to make a deepcopy of the args object before calling optuna_override_undefined