Update optuna_objective.py
Local step now represents the number of (locally) seen images, rather an iteration counter for batches. It makes more sense to specify the frequency of metric computation based on the number of seen images, especially if tuning e.g. the batch size, so that the frequency is reasonably similar in 'time', even with various batch sizes