Update loss.py
Added the angle and ecal loss terms from the custom anglepgan loss functions to the wgan loss function. Need to check if the reduce_mean() calls are in the correct spot (see the comments); if weighting the loss terms is necessary; and potentially adding the other loss terms gul dropped but sofia likes.